Update test to expect CPU Usage as a Nullable Float

+++ tests/glesys/requests/compute/helper.rb
@@ -114,7 +114,7 @@ class Glesys
               'unit'  => Fog::Nullable::String
             },
             'cpu'       => {
-              'usage' => Fog::Nullable::Integer,
+              'usage' => Fog::Nullable::Float,
               'max'   => Fog::Nullable::Integer,
               'unit'  => Fog::Nullable::String
             },
